requirements.txt
ansicolors==1.1.8
distro==1.5.0
pyfiglet==0.7
importlib_resources==5.4.0
dataclasses==0.8; python_version < '3.7'
types-dataclasses==0.1.5; python_version < '3.7'
packaging==20.9
setuptools==56.2.0
requests==2.26.0
semver==2.13.0
python-hosts==1.0.1
tabulate==0.8.9
types-tabulate==0.8.2
PyYAML==5.4.1
types-PyYAML==5.4.10
pyobjc-core==8.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==8.1; sys_platform == 'darwin'
requests-unixsocket==0.2.0
dpath==2.0.5
